funct: Creating components for the map list  panel
