+++ src/components/ListMaps/index.css
@@ -68,7 +68,6 @@
   color: white;
   font-size: 28px;
   border: 1px solid white;
-  opacity: .5;
   margin-bottom: 16px;
 }
 
@@ -78,14 +77,14 @@
 }
 
 .btnEnabled:hover {
-  opacity: 1;
+  background-color: #2c3e50;
   transition: .04s all ease-in-out;
   cursor: pointer;
 }
 
 .btnMap:hover {
   background-color: #2c3e50;
-  transition: .18s all ease-in-out;
+  transition: .04s all ease-in-out;
 }
 
 .mapName {
